Refactor the protobuf objects names based on the new release 1.3.1
